requirements.txt
pandas
numpy
matplotlib
seaborn
python-dotenv
langchain-openai
deepeval
